core/requirements.txt
BeautifulSoup==3.2.1
Django==1.6.2
MySQL-python
South==0.8.4
argparse==1.2.1
distribute
django-debug-toolbar==0.11.0
django-ckeditor-updated==4.4.0
django-filter==0.7
django-mptt==0.6.0
django-redis==3.5.1
django-redis-sessions==0.4.0
lxml==3.3.1
python-dateutil==2.1
python-logstash==0.2.1
redis==2.9.1
six==1.4.1
sqlparse==0.1.10
suds==0.4
celery==3.1.12
django-celery==3.1.10
uuid==1.30
wsgiref==0.1.2

